requirements.txt
astroid==2.3.1
beautifulsoup4==4.8.1
Django==2.2.6
django-cors-headers==3.1.1
djangorestframework==3.10.3
isort==4.3.21
lazy-object-proxy==1.4.2
lxml==4.4.1
mccabe==0.6.1
Pillow==6.2.0
pkg-resources==0.0.0
pylint==2.4.2
pytz==2019.3
selenium==3.141.0
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.6
wrapt==1.11.2
